+++ R/data.R
@@ -46,6 +46,30 @@ NULL
 #' @keywords data
 NULL
 
+#' Seed geographical dictionary in Italian
+#'
+#' @name data_dictionary_newsmap_it
+#' @docType data
+#' @author Giuseppe Carteny \email{giuseppe.carteny@unimi.it}
+#' @keywords data
+NULL
+
+#' Seed geographical dictionary in Hebrew
+#'
+#' @name data_dictionary_newsmap_he
+#' @docType data
+#' @author Elad Segev \email{eladseg@gmail.com}
+#' @keywords data
+NULL
+
+#' Seed geographical dictionary in Arabic
+#'
+#' @name data_dictionary_newsmap_ar
+#' @docType data
+#' @author Dai Yamao \email{daiyamao@scs.kyushu-u.ac.jp}
+#' @keywords data
+NULL
+
 #' Seed geographical dictionary in Chinese (simplified)
 #'
 #' @name data_dictionary_newsmap_zh_cn