+++ R/textmodel.R
@@ -8,12 +8,12 @@
 #' @param x dfm from which features will be extracted
 #' @param y dfm in which features will be class labels
 #' @param smooth smoothing parameter for word frequency
-#' @param verbose if \code{TRUE}, show progress of training
-#' @import quanteda
+#' @param verbose if `TRUE`, show progress of training
+#' @importFrom quanteda is.dfm dfm_trim nfeat
 #' @references Kohei Watanabe. 2018.
-#'   "\href{http://www.tandfonline.com/eprint/dDeyUTBrhxBSSkHPn5uB/full}{Newsmap:
-#'    semi-supervised approach to geographical news classification.}"
-#'   \emph{Digital Journalism} 6(3): 294-309.
+#'   "[Newsmap:
+#'    semi-supervised approach to geographical news classification.](https://www.tandfonline.com/eprint/dDeyUTBrhxBSSkHPn5uB/full)"
+#'   *Digital Journalism* 6(3): 294-309.
 #' @export
 #' @examples
 #' require(quanteda)
@@ -35,7 +35,6 @@ textmodel_newsmap <- function(x, y, smooth = 1, verbose = quanteda_options('verb
     if (!is.dfm(x) || !is.dfm(y))
         stop('x and y have to be dfm')
 
-    docvars(x) <- NULL
     x <- dfm_trim(x, min_termfreq = 1)
     y <- dfm_trim(y, min_termfreq = 1)
 
@@ -48,18 +47,21 @@ textmodel_newsmap <- function(x, y, smooth = 1, verbose = quanteda_options('verb
                     dimnames = list(colnames(y), colnames(x)))
     if (verbose)
         cat("Training for class: ")
+    m <- colSums(x)
     for (key in sort(featnames(y))) {
         if (verbose)
             cat(key, " ", sep = "")
-        temp <- dfm_group(x, factor(as.vector(y[,key]) > 0,
-                                    levels = c('TRUE', 'FALSE')), fill = TRUE)
-        temp <- temp + smooth
-        temp <- temp / rowSums(temp) # likelihood
-        model[key,] <- as.vector(log(temp['TRUE',]) - log(temp['FALSE',])) # likelihood-ratio
+        s <- colSums(x[as.logical(y[,key] > 0),])
+        v1 <- s + smooth
+        v0 <- m - s + smooth
+        model[key,] <- log(v1 / sum(v1)) - log(v0 / sum(v0)) # log-likelihood ratio
     }
     if (verbose)
         cat("\n")
-    result <- list(model = model, data = x, feature = colnames(model))
+    result <- list(model = model,
+                   data = x,
+                   feature = colnames(model),
+                   call = match.call())
     class(result) <- "textmodel_newsmap"
     return(result)
 }
@@ -69,15 +71,16 @@ textmodel_newsmap <- function(x, y, smooth = 1, verbose = quanteda_options('verb
 #' Predict document class using trained a Newsmap model
 #' @param object a fitted Newsmap textmodel
 #' @param newdata dfm on which prediction should be made
-#' @param confidence.fit if \code{TRUE}, likelihood ratio score will be returned
-#' @param rank rank of class to be predicted. Only used when \code{type = "top"}.
-#' @param type if \code{top}, return the most likely class specified by
-#'   \code{rank}; otherswise return a matrix of lilelyhood ratio scores for all
+#' @param confidence.fit if `TRUE`, likelihood ratio score will be returned
+#' @param rank rank of class to be predicted. Only used when `type = "top"`.
+#' @param type if `top`, returns the most likely class specified by
+#'   `rank`; otherwise return a matrix of likelihood ratio scores for all
 #'   possible classes
 #' @param ... not used.
 #' @method predict textmodel_newsmap
 #' @export
-#' @import quanteda methods
+#' @importFrom methods as
+#' @importFrom quanteda dfm_match dfm_weight docnames featnames quanteda_options
 predict.textmodel_newsmap <- function(object, newdata = NULL, confidence.fit = FALSE, rank = 1L,
                                       type = c("top", "all"), ...) {
 
@@ -100,22 +103,24 @@ predict.textmodel_newsmap <- function(object, newdata = NULL, confidence.fit = F
     if (type == 'top') {
         if (confidence.fit) {
             if (ncol(temp)) {
-                result <- list(class = apply(temp, 1, function(x) names(get_nth(x, rank))),
-                               confidence.fit = unname(apply(temp, 1, function(x) get_nth(x, rank))))
+                result <- list(class = get_nth(temp, rank, "class"),
+                               confidence.fit = unname(get_nth(temp, rank, "conf")))
             } else {
                 result$class <- rep(NA, nrow(temp))
             }
             result$class[is_empty] <- NA
             result$confidence.fit[is_empty] <- NA
             names(result$class) <- docnames(data)
+            result$class <- factor(result$class, levels = rownames(model))
         } else {
             if (ncol(temp)) {
-                result <- apply(temp, 1, function(x) names(get_nth(x, rank)))
+                result <- get_nth(temp, rank, "class")
             } else {
                 result <- rep(NA, nrow(temp))
             }
             result[is_empty] <- NA
             names(result) <- docnames(data)
+            result <- factor(result, levels = rownames(model))
         }
     } else {
         result <- temp
@@ -125,19 +130,32 @@ predict.textmodel_newsmap <- function(object, newdata = NULL, confidence.fit = F
     return(result)
 }
 
-get_nth <- function(x, rank) {
-    sort(x, decreasing = TRUE)[rank]
+get_nth <- function(x, rank, type = c("class", "conf")) {
+
+    type <- match.arg(type)
+    for (i in seq_len(rank - 1)) {
+        x <- replace(x, cbind(seq_len(nrow(x)), max.col(x)), -Inf)
+    }
+    if (type == "conf") {
+        result <- x[cbind(seq_len(nrow(x)), max.col(x))]
+    } else {
+        result <- colnames(x)[max.col(x)]
+    }
+    names(result) <- rownames(x)
+    return(result)
 }
 
 #' @noRd
 #' @method summary textmodel_newsmap
 #' @export
 summary.textmodel_newsmap <- function(object, n = 10, ...) {
-    result <- list(classes = utils::head(rownames(object$model), n),
-                   features = utils::head(colnames(object$model), n),
-                   documents = utils::head(rownames(object$data), n))
-    class(result) <- 'textmodel_newsmap_summary'
-    return(result)
+    result <- list(
+        "call" = object$call,
+        "labels" = rownames(object$model)
+    )
+    if (!is.null(object$data))
+        result$data.dimension <- dim(object$data)
+    as.summary.textmodel(result)
 }
 
 #' @noRd
@@ -183,7 +201,7 @@ print.textmodel_newsmap_summary <- function(x, ...) {
 # positive (fp), true negative (tn) and false negative (fn) cases for each
 # predicted class. It also calculates precision, recall and F1 score based on
 # these counts.
-#' @param x vercor of predicted classes
+#' @param x vector of predicted classes
 #' @param y vector of true classes
 #' @export
 #' @examples
@@ -214,11 +232,11 @@ accuracy <- function(x, y) {
     return(result)
 }
 
-#' Calcualte micro and macro average measures of accuracy
+#' Calculate micro and macro average measures of accuracy
 #'
-#' This function calculates micro-averave precision (p) and recall (r) and
+#' This function calculates micro-average precision (p) and recall (r) and
 #' macro-average precision (P) and recall (R) based on a confusion matrix from
-#' \code{accuracy()}.
+#' `accuracy()`.
 #' @param object output of accuracy()
 #' @param ... not used.
 #' @method summary textmodel_newsmap_accuracy
@@ -238,16 +256,22 @@ summary.textmodel_newsmap_accuracy <- function(object, ...) {
     return(result)
 }
 
-#' Compute average feature entroy
+#' Compute average feature entropy (AFE)
+#'
+#' AFE computes randomness of occurrences features in labelled documents.
 #' @param x a dfm for features
 #' @param y a dfm for labels
 #' @param smooth a numeric value for smoothing to include all the features
+#' @importFrom quanteda.textstats textstat_entropy
+#' @importFrom quanteda is.dfm nfeat featnames colSums rowSums dfm_subset as.dfm
 #' @export
 afe <- function(x, y, smooth = 1) {
     if (!is.dfm(x) || !is.dfm(y))
         stop('x and y have to be dfm')
     e <- textstat_entropy(group_topics(x, y) + smooth,
                           margin = "features")
+    if (is.data.frame(e))
+        e <- e$entropy
     return(mean(e))
 }
 
@@ -255,9 +279,7 @@ group_topics <- function(x, y) {
     result <- matrix(NA, nrow = nfeat(y), ncol = nfeat(x),
                      dimnames = list(featnames(y), featnames(x)))
     for (i in seq_len(nfeat(y))) {
-        result[i,] <- colSums(dfm_subset(x, rowSums(y[,i]) > 0))
+        result[i, ] <- colSums(dfm_subset(x, rowSums(y[ ,i]) > 0))
     }
     return(as.dfm(result))
 }
-
-