+++ R/utils.R
@@ -1,0 +1,8 @@
+
+#' @export
+#' @method print textmodel_newsmap
+print.textmodel_newsmap <- function(x, ...) {
+    cat("\nCall:\n")
+    print(x$call)
+    cat("\n")
+}