+++ NEWS.md
@@ -1,3 +1,35 @@
+
+* Add `print()` methods for fitted models
+
+# Changens in v0.7.4
+
+* Change predicted values to factor with all labels in training data
+
+# Changes in v0.7.3
+
+* Update tests for quanteda v3.0
+
+# Changes in v0.7.2
+
+* Make `predict()` significantly faster
+
+# Changes in v0.7.1
+
+* Improve efficiency of `textmodel_newsmap()`
+* Add compatibility with newer `textstat_entropy()`
+
+# Changes in v0.7.0
+
+* Add Hebrew and Arabic seed dictionaries
+
+# Changes in v0.6.9
+
+* Clean up Italian, German and Spanish seed dictionaries
+
+# Changes in v0.6.8
+
+* Add Italian seed dictionary
+
 # Changes in v0.6.7
 
 ## Dictionary updates
@@ -24,14 +56,14 @@
 
 ## Bug fixes
 
-* Fix error in textmodel_newsmap() when smooth is < 1.0
-* Fitted models no longer include classes that did not occure in training set
+* Fix error in `textmodel_newsmap()` when smooth is < 1.0
+* Fitted models no longer include classes that did not occur in training set
 
 # Changes in v0.4.6
 
 ## New functions
 
-* Add coef() and coefficients() methods
+* Add `coef()` and `coefficients()` methods
 
 ## Dictionary updates